+++ src/components/pages/NotFound.vue
@@ -1,0 +1,13 @@
+<template>
+    <div>
+Not found
+    </div>
+</template>
+
+<script setup>
+
+</script>
+
+<style lang="scss" scoped>
+
+</style>
