+++ src/components/pages/PasswordReset.vue
@@ -1,0 +1,13 @@
+<template>
+    <div>
+        ResetPAssword
+    </div>
+</template>
+
+<script setup>
+
+</script>
+
+<style lang="scss" scoped>
+
+</style>
